+++ packages/network/lib/network.dart
@@ -1,4 +1,6 @@
 export 'src/exception.dart';
 export 'src/response.dart';
-export 'src/get.dart';
-export 'src/post.dart';
+export 'src/methods/get.dart';
+export 'src/methods/post.dart';
+export 'src/methods/put.dart';
+export 'src/methods/delete.dart';