+++ packages/network/lib/src/methods/delete.dart
@@ -1,0 +1,20 @@
+import 'dart:async' show Future;
+import 'package:http/http.dart' as http show delete, Response;
+import 'package:network/src/exception.dart';
+import 'package:network/src/response.dart';
+import 'package:network/src/utils.dart';
+
+Future<T> delete<T extends BinaryResponse>(String url,
+    {Map<String, String> headers}) async {
+  final http.Response httpResponse = await http.delete(url, headers: headers);
+
+  final int statusCode = httpResponse.statusCode;
+
+  final response = makeResponseByType<T>(statusCode, httpResponse.bodyBytes);
+
+  if (statusCode < 200 || statusCode >= 400) {
+    throw NetworkException<T>(response);
+  }
+
+  return response;
+}