+++ packages/network/CHANGELOG.md
@@ -3,6 +3,11 @@
 [comment]: <> (### Breaking Changes or ### New Features)
 [comment]: <> (* Change description)
 
+## 0.0.2
+
+* Removed debug print
+* Added hooks for `delete` & `put` http methods
+
 ## 0.0.1
 
 * Initial release